Update mesh-details.groovy
I'm running hubitat platform 2.3.8.140 , C8-pro and device that are connected directly to HUB have neoghborID 3 5 7 , which are part of the hub but are not present in the list of devices. As the result findDeviceByDecId(neighborId) returns NUL in displayRowDetail() function call and the device details page does not show up at all.