Update mesh-details.groovy
Hi Tony.  If you think it is possible device IDs can be lower than 10, what about this change,to check if the neighbor device ID exists in the list of devices before adding it to the neighbors map. With the change you are suggesting, device IDs 3,5,7 are still present in the neighbors map and show up in device details as unknown devices. But I don't think people would like to see that, the unknown device is actually a hub, not an unknown device. The presence of unknown devices is misleading, it creates an impression that something is wrong, and zwave list of neighbors  is corrupted and contains device IDs that are not valid.