Update mesh-details.groovy
If you think actual devices can have an ID less than 10, what about this change, to check if the neighbor ID exists in the list of devices before adding it to the neighbors list. With the change you are suggesting, device IDs 3,5,7 are still present in the neighborsMap and show up as unknown devices in the device details. But I think this is misleading, it creates an impression that device's zwave table is corrupted and contains invalid device IDs. While it is not true, 3,5,7 are actually hub IDs.